Rename Prayer Flicking plugin to Prayer
